mathiaswo/requirements.txt
black
requests
python-dotenv
